post delete done, view all types of users with scroll done

+++ php/auth.php
@@ -104,7 +104,10 @@ class auth{
             $file = $row['photoPath1'];
             $singl = explode(',',$file);
             foreach($singl as $s){
-                unlink('.'.$s);
+                if($s){
+                    unlink('.'.$s);
+                }
+                
             }
             
         }
